// src/enums/globaEnums.ts
export enum State {
    invalid = 'invalid',
    unsuccess = 'unsuccess'
}

export enum Branch {
    selectBranch = 0,
    menBranch = 1,
    womenBranch = 2
}

export enum Category {
    selectCategory = 0,
    r = 1,
    e = 2,
    d = 3,
    c = 4,
    b = 5,
    a = 6,
}

export enum Role {
    selectRole = 0,
    admin= 1,
    manager = 2,
    coach = 3
}